Update Server.php
lines 60/63 - Pingback_Utility::isURL() returns true when a url is valid, so a fault is set when there is a valid url. I added ! in front of the method

line 69 - Pingback_Utility::isBacklinking() - same deal. It returns true when there is a backlink which effectively causes the fault on the same line. I added ! in front of the call

line 119 - changed to return $this->_requestSource; because that's where the source url is actually stored

line 124 - changed to return $this->_requestTarget; because that's where the target url is actually stored